Remove unused npm dependencies
It looks like these dependencies aren't actually used anywhere, so stop depending on them.

The motivation for this change is that the dependency on `resolve` causes errors to be logged in logs during plugin discovery, because the package includes tests that have malformed `package.json` files. If we don't actually need the dependency it'd be nice to remove it so those errors are no longer logged (so users don't ask us about them).

Also remove `read-package-tree` since it is deprecated.

Also remove `builtin-modules` since it doesn't appear to be used either.